# speed up (de)serialization of Base bits types in abstract containers