helps #30148 

Before:
Serialize array of `Union{Missing,Int}`:
```
  7.327029 seconds (100.00 M allocations: 1.990 GiB, 4.15% gc time)
```
Deserialize that:
```
 54.608122 seconds (400.37 M allocations: 8.810 GiB, 1.18% gc time)
```
Serialize array of `Union{Missing,UInt}`:
```
 22.448258 seconds (200.19 M allocations: 3.493 GiB, 3.03% gc time)
```
Deserialize that:
```
 62.783948 seconds (399.90 M allocations: 8.787 GiB, 0.94% gc time)
```

Same tests with this PR:
```
  5.504755 seconds (100.00 M allocations: 1.990 GiB, 12.72% gc time)
  6.396924 seconds (199.90 M allocations: 4.317 GiB, 5.81% gc time)
  4.939101 seconds (100.22 M allocations: 2.001 GiB, 8.17% gc time)
  6.633005 seconds (200.56 M allocations: 4.349 GiB, 4.78% gc time)
```

The times are a bit variable but the basic story is clear. I tested both Int and UInt, since Int already had some special-case code. Other Base number types (Int8, UInt8, ...) should have roughly the same performance characteristics as UInt.