@KristofferC 2018-11-30
I thought we gave no guarantees about being able to deserialize data from any previous version of Julia?