@JeffBezanson 2018-12-01
No this should be 100% backwards compatible. It just uses if statements instead of dispatch, and additional specialized methods. There were already special tags for those types.